+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added async support for Tokio runtime.
+- Added `doc_auto_cfg` feature.
+
 ### Fixed
 
 - Added missing method comments to improve documentation clarity.